상세 페이지 컴포넌트 분리, 글 작성 및 수정 페이지에서 글 제목 삭제
